import calendar data from the service
